adjust action api post property (insert/edit)

+++ api/v1.0/api_property.php
@@ -26,11 +26,10 @@ if(isset($_GET['action'])){
 			require_once($global['root-url-model'].$version."/User.php");
 			$obj_user = new User();
 
-			//===================================== insert property ========================================
-			//start insert property
-			if($_GET['action'] == 'insert_data' && isset($_REQUEST['user_id']) && isset($_REQUEST['auth_token'])){
+			//===================================== post ========================================
+			//start post
+			if($_GET['action'] == 'post' && isset($_REQUEST['user_id']) && isset($_REQUEST['auth_token'])){
 				$obj_connect->up();
-				$R_message = array("status" => "400", "message" => "Insert Data Failed");
 				$N_user_id = mysql_real_escape_string($_REQUEST['user_id']);
 				$N_auth_token = mysql_real_escape_string($_REQUEST['auth_token']);
 
@@ -66,49 +65,62 @@ if(isset($_GET['action'])){
 				$N_create_date = mysql_real_escape_string($_REQUEST['create_date']);
 
 				if($obj_user->check_code($N_auth_token, $N_user_id)){//check code
-					$result = $obj_property->insert_data($N_token, $N_owner, $N_title, $N_hak, $N_province, $N_city, $N_kecamatan, $N_kelurahan, $N_address, $N_zip, $N_type, $N_status_property, $N_sertifikat, $N_menghadap, $N_lebar_depan, $N_panjang_tanah, $N_luas_tanah, $N_luas_bangunan, $N_bed, $N_bed_plus, $N_bath, $N_floor, $N_luas_apartment, $N_description, $N_hashtag, $N_price, $N_komisi, $N_status, $N_create_date);
-					//var_dump($result);
-					if($result == 1){
-						//insert image
-						require_once($global['root-url']."packages/SimpleImage.php"); // class simple image
-						if(isset($_FILES['image']['name'])){
-							for($i = 0; $i < count($_FILES['image']['name']); $i++){
-								if(!empty($_FILES['image']['name'][$i])){
-									$N_image_token = mysql_real_escape_string($_REQUEST['image_token'][$i]);
-									$allowed_ext = array('jpg', 'jpeg', 'png', 'gif');
-									$file_name = cleanSpace($_FILES['image']['name'][$i]);
-									$file_ext = strtolower(end(explode('.', $file_name)));
-									$file_size = $_FILES['image']['size'][$i];
-									$file_tmp = $_FILES['image']['tmp_name'][$i];
-									$ran = rand();
-						    		$timestamp = time();
-									
-									if(in_array($file_ext, $allowed_ext) === true){
-										if($file_size < 10044070){
-											//save image in server
-											$file_loc = $global['root-url']."uploads/property/".$timestamp.$ran.$file_name; 
-											$file_locThmb = $global['root-url']."uploads/property-thmb/".$timestamp.$ran.$file_name;
+					$check_exist = $obj_property->check_exist($N_token);
+					//var_dump($check_exist);
+					if($check_exist == 0){
+						$result = $obj_property->insert_data($N_token, $N_owner, $N_title, $N_hak, $N_province, $N_city, $N_kecamatan, $N_kelurahan, $N_address, $N_zip, $N_type, $N_status_property, $N_sertifikat, $N_menghadap, $N_lebar_depan, $N_panjang_tanah, $N_luas_tanah, $N_luas_bangunan, $N_bed, $N_bed_plus, $N_bath, $N_floor, $N_luas_apartment, $N_description, $N_hashtag, $N_price, $N_komisi, $N_status, $N_create_date);
+						//var_dump($result);
+						if($result == 1){
+							//insert image
+							require_once($global['root-url']."packages/SimpleImage.php"); // class simple image
+							if(isset($_FILES['image']['name'])){
+								for($i = 0; $i < count($_FILES['image']['name']); $i++){
+									if(!empty($_FILES['image']['name'][$i])){
+										$N_image_token = mysql_real_escape_string($_REQUEST['image_token'][$i]);
+										$allowed_ext = array('jpg', 'jpeg', 'png', 'gif');
+										$file_name = cleanSpace($_FILES['image']['name'][$i]);
+										$file_ext = strtolower(end(explode('.', $file_name)));
+										$file_size = $_FILES['image']['size'][$i];
+										$file_tmp = $_FILES['image']['tmp_name'][$i];
+										$ran = rand();
+							    		$timestamp = time();
+										
+										if(in_array($file_ext, $allowed_ext) === true){
+											if($file_size < 10044070){
+												//save image in server
+												$file_loc = $global['root-url']."uploads/property/".$timestamp.$ran.$file_name; 
+												$file_locThmb = $global['root-url']."uploads/property-thmb/".$timestamp.$ran.$file_name;
 
-											//save image in database
-											$file_loc1 = "uploads/property/".$timestamp.$ran.$file_name; 
-											$file_locThmb1 = "uploads/property-thmb/".$timestamp.$ran.$file_name;
+												//save image in database
+												$file_loc1 = "uploads/property/".$timestamp.$ran.$file_name; 
+												$file_locThmb1 = "uploads/property-thmb/".$timestamp.$ran.$file_name;
 
-											if(move_uploaded_file($file_tmp, $file_loc))
-											{
-												$image = new SimpleImage();
-												$image->load($file_loc);
-												$image->resize(200,200);
-												$image->save($file_locThmb);
+												if(move_uploaded_file($file_tmp, $file_loc))
+												{
+													$image = new SimpleImage();
+													$image->load($file_loc);
+													$image->resize(200,200);
+													$image->save($file_locThmb);
+												}
+												
+												$obj_image->insert_data($N_image_token, $N_token, $file_loc1, $file_locThmb1);
 											}
-											
-											$obj_image->insert_data($N_image_token, $N_token, $file_loc1, $file_locThmb1);
 										}
 									}
 								}
 							}
+							$R_message = array("status" => "201", "message" => "Insert Data Success");
+						}else{
+							$R_message = array("status" => "400", "message" => "Insert Data Failed");
 						}
-
-						$R_message = array("status" => "200", "message" => "Insert Data Success");
+					}else{
+						$result = $obj_property->update_data($N_token, $N_owner, $N_title, $N_hak, $N_province, $N_city, $N_kecamatan, $N_kelurahan, $N_address, $N_zip, $N_type, $N_status_property, $N_sertifikat, $N_menghadap, $N_lebar_depan, $N_panjang_tanah, $N_luas_tanah, $N_luas_bangunan, $N_bed, $N_bed_plus, $N_bath, $N_floor, $N_luas_apartment, $N_description, $N_hashtag, $N_price, $N_komisi);
+						//var_dump($result);
+						if($result == 1){
+							$R_message = array("status" => "200", "message" => "Update Data Success");
+						}else{
+							$R_message = array("status" => "400", "message" => "Update Data Failed");
+						}
 					}
 				}//check code
 				else{
@@ -117,7 +129,7 @@ if(isset($_GET['action'])){
 
 				$obj_connect->down();
 				echo json_encode($R_message);	
-			}//end insert property
+			}//end post
 
 			//===================================== sync ========================================
 			//start sync
